Boston Housing Data
The Boston housing dataset is a dataset that has median value of the house along with 13 other parameters that could potentially be related to housing prices. There are 506 observations in the data for 14 variables including the median price of house in Boston. There are 12 numerical variables in our dataset and 1 categorical variable. The goal of this project is to build best predictive model to estimate the median price of owner-occupied homes in Boston.